src/templates/packages/o-auth/src/@api/o-auth/client/resolvers/o-auth-update-client.resolver.ts
import { UseGuards } from '@nestjs/common';
import { Resolver, Args, Mutation } from '@nestjs/graphql';
import { Constraint, QueryStatement, Timezone } from 'aurora-ts-core';

// authorization
import { Permissions } from '../../../../@api/iam/shared/decorators/permissions.decorator';
import { AuthenticationJwtGuard } from '../../../../@api/o-auth/shared/guards/authentication-jwt.guard';
import { AuthorizationGuard } from '../../../../@api/iam/shared/guards/authorization.guard';

// @apps
import { OAuthUpdateClientHandler } from '../handlers/o-auth-update-client.handler';
import { OAuthClient, OAuthUpdateClientInput } from '../../../../../graphql';

@Resolver()
@Permissions('oAuth.client.update')
@UseGuards(AuthenticationJwtGuard, AuthorizationGuard)
export class OAuthUpdateClientResolver
{
    constructor(
        private readonly handler: OAuthUpdateClientHandler,
    ) {}

    @Mutation('oAuthUpdateClient')
    async main(
        @Args('payload') payload: OAuthUpdateClientInput,
        @Constraint() constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<OAuthClient>
    {
        return await this.handler.main(
            payload,
            constraint,
            timezone,
        );
    }
}